requirements.txt
aiogram==3.17.0
google-generativeai==0.8.3
pyyaml==6.0.2
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic==2.10.0
pydantic-settings==2.7.0